Menu li items set to justify arrangement
Per Toni's specs, menu items now line up so that the far right and left
link are right up against the edge. Also, I've pulled in the logo as it
was out of line and now that lines up correctly with the main nav menu.
We may have to adjust other  blocks/elements in the site to say within
the main container now.