Refactor default page and fix page gap bug
Using the default template on the homepage results in a gap between the
header of the page and the block being used. This fix removes the gap
and puts the extra spacing on only non-homepage headers.

I also refactored page.php so that it is a bit more clear how the
special category components load in.