Add meta description tag, that links to the site tagline
Because the cookie banner comes naer the top of the DOM (for accessibility reasons), that text is currently being pulled out by google and shown in search results. This adds in a meta tag that pulls in the site tagline as the site description. This means this should show on google instead of the cookie info.

I've also updated the meta tag with the theme name in to say hale, instead of nightingale.